[OPIK-1160] refactor
