+++ apps/opik-frontend/src/components/shared/ChartHorizontalLegendContent/ChartHorizontalLegendContent.tsx
@@ -2,15 +2,21 @@ import React from "react";
 import * as RechartsPrimitive from "recharts";
 import { OnChangeFn } from "@/types/shared";
 import TooltipWrapper from "@/components/shared/TooltipWrapper/TooltipWrapper";
+import { cn } from "@/lib/utils";
 
-const ExperimentRadarChartLegendContent = React.forwardRef<
+type ChartHorizontalLegendContentProps = React.ComponentProps<
+  typeof RechartsPrimitive.Legend
+> &
+  React.ComponentProps<"div"> & {
+    setActiveLine: OnChangeFn<string | null>;
+    chartId: string;
+    containerClassName?: string;
+  };
+
+const ChartHorizontalLegendContent = React.forwardRef<
   HTMLDivElement,
-  React.ComponentProps<typeof RechartsPrimitive.Legend> &
-    React.ComponentProps<"div"> & {
-      setActiveLine: OnChangeFn<string | null>;
-      chartId: string;
-    }
->(({ payload, color, setActiveLine }, ref) => {
+  ChartHorizontalLegendContentProps
+>(({ payload, color, setActiveLine, containerClassName }, ref) => {
   const handleMouseEnter = (id: string) => {
     setActiveLine(id);
   };
@@ -26,10 +32,17 @@ const ExperimentRadarChartLegendContent = React.forwardRef<
   return (
     <div
       ref={ref}
-      className="mt-6 flex size-full max-h-[34px] justify-center overflow-auto"
+      className={cn(
+        "-mt-2.5 w-full max-w-full pt-6 text-center",
+        containerClassName,
+      )}
       onMouseLeave={handleMouseLeave}
     >
-      <div className="group inline-flex max-w-full flex-wrap items-center justify-center space-x-2">
+      <div
+        className={
+          "group inline-flex max-w-full flex-wrap items-center justify-center space-x-2"
+        }
+      >
         {payload.map((item, idx) => {
           const key = `${item.value || "value"}`;
           const indicatorColor = color || item.color;
@@ -37,7 +50,7 @@ const ExperimentRadarChartLegendContent = React.forwardRef<
           return (
             <div
               key={key + idx}
-              className="relative min-w-0 cursor-pointer pl-3 duration-200 group-hover-except-self:opacity-60"
+              className="relative min-w-0 pl-3 duration-200 group-hover-except-self:opacity-60"
               onMouseEnter={() => handleMouseEnter(item.value)}
             >
               <TooltipWrapper content={item.value}>
@@ -61,7 +74,6 @@ const ExperimentRadarChartLegendContent = React.forwardRef<
     </div>
   );
 });
-ExperimentRadarChartLegendContent.displayName =
-  "ExperimentRadarChartLegendContent";
+ChartHorizontalLegendContent.displayName = "ChartHorizontalLegendContent";
 
-export default ExperimentRadarChartLegendContent;
+export default ChartHorizontalLegendContent;
